+++ enclone_print/src/lib.rs
@@ -8,3 +8,4 @@ pub mod print_utils2;
 pub mod print_utils3;
 pub mod print_utils4;
 pub mod print_utils5;
+pub mod proc_lvar;
